Working 'How to Play' reacts to click on page. Reactified HowToPlay and Instructions.
