+++ src/index.html
@@ -35,9 +35,6 @@
       </section>
     </footer>
 </div>
+<script type="text/javascript" src="bundle.js"></script>
 </body>
 </html>
-
-<script type="text/javascript" src="bundle.js"></script>
-<script>
-</script>
